expertoption,
As CSO of the Human Rights Foundation, Gladstein is uniquely positioned to detail the rise of Bitcoin from cypherpunk dream to the real-life Bitcoin stories happening to real people across the globe. For people around the world, outside of Wall Street, Bitcoin offers a means of freedom from inflation, political strife, and an outdated monetary system. For these people, the majority of the world’s population, it might even save their lives.